Renamed DoNotUseLabels to DoNotUseGoto
DoNotUseLabels analyzed Goto and Labels, so combined into DoNotUseGoto
