+++ src/Analyzer/Factory.php
@@ -52,7 +52,6 @@ class Factory
                 new AnalyzerPass\Statement\OldConstructor(),
                 new AnalyzerPass\Statement\ConstantNaming(),
                 new AnalyzerPass\Statement\DoNotUseInlineHTML(),
-                new AnalyzerPass\Statement\DoNotUseLabels(),
             ]
         );
         $analyzer->bind();
