+++ src/Analyzer/Pass/Statement/DoNotUseGoto.php
@@ -4,6 +4,8 @@ namespace PHPSA\Analyzer\Pass\Statement;
 
 use PhpParser\Node\Stmt;
 use PhpParser\Node;
+use PhpParser\Node\Stmt\Goto_;
+use PhpParser\Node\Stmt\Label;
 use PHPSA\Analyzer\Pass;
 use PHPSA\Context;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
@@ -11,15 +13,28 @@ use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 class DoNotUseGoto implements Pass\ConfigurablePassInterface, Pass\AnalyzerPassInterface
 {
     /**
-     * @param Stmt\Goto_ $stmt
+     * @param $stmt
      * @param Context $context
      * @return bool
      */
-    public function pass(Stmt\Goto_ $stmt, Context $context)
+    public function pass($stmt, Context $context)
     {
-        $context->notice('do_not_use_goto', 'Do not use goto statements', $stmt);
-
-        return true;
+        if ($stmt instanceof Label) {
+            $context->notice(
+                'do_not_use_goto',
+                'Do not use labels',
+                $stmt
+            );
+            return true;
+        } elseif ($stmt instanceof Goto_) {
+            $context->notice(
+                'do_not_use_goto',
+                'Do not use goto statements',
+                $stmt
+            );
+            return true;
+        }
+        return false;
     }
 
     /**
@@ -29,8 +44,7 @@ class DoNotUseGoto implements Pass\ConfigurablePassInterface, Pass\AnalyzerPassI
     {
         $treeBuilder = new TreeBuilder();
         $treeBuilder->root('do_not_use_goto')
-            ->canBeDisabled()
-        ;
+            ->canBeDisabled();
 
         return $treeBuilder;
     }
@@ -41,7 +55,8 @@ class DoNotUseGoto implements Pass\ConfigurablePassInterface, Pass\AnalyzerPassI
     public function getRegister()
     {
         return [
-            Stmt\Goto_::class,
+            Goto_::class,
+            Label::class,
         ];
     }
 }
