+++ tests/analyze-fixtures/Statement/DoNotUseGoto.php
@@ -2,7 +2,8 @@
 
 namespace Tests\Analyze\Fixtures\Statement;
 
-class DoNotUseGoto {
+class DoNotUseGoto
+{
     public function testGoto()
     {
         marker:
@@ -11,7 +12,19 @@ class DoNotUseGoto {
 
     public function testNoGoto()
     {
+        $a = 1;
+    }
 
+    public function testLabel()
+    {
+        marker1:
+        $a = 1;
+
+    }
+
+    public function testNoLabel()
+    {
+        $a = 2;
     }
 }
 
@@ -20,8 +33,20 @@ class DoNotUseGoto {
 [
     {
         "type": "do_not_use_goto",
-        "message": "Do not use goto statements",
+        "message": "Do not use labels",
         "file": "DoNotUseGoto.php",
         "line": 8
+    },
+    {
+        "type": "do_not_use_goto",
+        "message": "Do not use goto statements",
+        "file": "DoNotUseGoto.php",
+        "line": 9
+    },
+    {
+        "type": "do_not_use_goto",
+        "message": "Do not use labels",
+        "file": "DoNotUseGoto.php",
+        "line": 19
     }
 ]
